Moving groups to new framework
